Add safe_add and saturating_add functions
Due to the interferences of reconfiguration agents,
signed integer overflows may occur during the SAG exploration:
in particular `max_int_value + 1` is common.

To deal with this, I made a `saturating_add(a, b)` function that
computes `a + b`, but will return `max_int_value` when computing
`a + b` would cause integer overflow.

Furthermore, there are some cases where integer overflow *should*
not happen, but it may happen anyway when I make mistakes.
To lighten the debugging burden of this, I also made a
`safe_add(a, b)` function that will compute `a + b`, but throws
an exception when this would lead to integer overflow.