Reconfiguration_agent.may_potentially_forbid_jobs
This commit adds a `may_potentially_forbid_jobs` method to
`Reconfiguration_agent`, which should return `true` if
the agent subclass also overrides the `is_allowed` method.

The exploration code needs to know this since some
optimizations rely on the assumption that jobs without
precedence constraints can certainly be scheduled when
they are released and a processor is available.